Relocate MTO Updater service object functions to move_task_order_updater file
